refactor Chart component to extract scale management and limit no-go zone logic into custom hooks useChartScale and useLimitNoGoZone, move shape location checkers import, fix swipe back cleanup by returning cleanup function, and remove keydown event listeners on drag end
